requirements.txt
anyio==3.6.2
certifi==2022.12.7
cfscrape==2.1.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
fastapi==0.94.1
flatbuffers==23.3.3
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
mpmath==1.3.0
numpy==1.24.2
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
packaging==23.0
protobuf==4.22.1
pydantic==1.10.6
python-multipart==0.0.6
qdrant-client==1.1.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
